Adds basic type support to the language
Adds basic type support for integers and booleans.

Changes the representation of types in the AST to use an enum
instead of a string. Updates the parser to handle type declarations and
expressions. This allows the compiler to perform type checking and
generate more efficient code.

Also renames "Function" to "fn" in the AST output for function
declarations and removes the "Statements:" prefix from statement node output.